bot/requirements.txt
nb-cli~=0.5.0
nonebot2~=2.0.0a14
nonebot-adapter-cqhttp~=2.0.0a14
pydantic~=1.8.2
alchemical==0.5.0